refactor: Remove the Pin2Color name as requested by @lucky01.
