+++ LibDmd/Converter/Vni/VniAnimationSet.cs
@@ -4,7 +4,7 @@ using System.Text;
 using LibDmd.Common;
 using LibDmd.Input.FileSystem;
 
-namespace LibDmd.Converter.Pin2Color
+namespace LibDmd.Converter.Vni
 {
 	public class VniAnimationSet : AnimationSet
 	{
@@ -31,14 +31,14 @@ namespace LibDmd.Converter.Pin2Color
 			var numAnimations = reader.ReadInt16BE();
 
 			if (Version >= 2) {
-				Logger.Trace("[pin2color] VNI[{1}] Skipping {0} bytes of animation indexes.", numAnimations * 4, reader.BaseStream.Position);
+				Logger.Trace("[vni] VNI[{1}] Skipping {0} bytes of animation indexes.", numAnimations * 4, reader.BaseStream.Position);
 				for (var i = 0; i < numAnimations; i++) {
 					reader.ReadUInt32();
 				}
 			}
 
 			Animations = new List<Animation>(numAnimations);
-			Logger.Debug("[pin2color] VNI[{3}] Reading {0} animations from {1} v{2}...", numAnimations, header, Version, reader.BaseStream.Position);
+			Logger.Debug("[vni] VNI[{3}] Reading {0} animations from {1} v{2}...", numAnimations, header, Version, reader.BaseStream.Position);
 
 			MaxWidth = 0;
 			MaxHeight = 0;
